requirements.txt
streamlit
pandas
mysql-connector-python
scikit-learn
plotly
numpy
openpyxl
reportlab
matplotlib
